refactor: extract host_creds into shared module for MLflow store classes
