+++ sagemaker_mlflow/host_creds.py
@@ -1,0 +1,56 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You
+# may not use this file except in compliance with the License. A copy of
+# the License is located at
+#
+#     http://aws.amazon.com/apache2.0/
+#
+# or in the "license" file accompanying this file. This file is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
+# ANY KIND, either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+
+import os
+
+from mlflow.utils import rest_utils
+
+from sagemaker_mlflow.mlflow_sagemaker_helpers import SageMakerMLflowHostMetadataProvider
+
+# Extra environment variables which take precedence for setting the basic/bearer
+# auth on http requests.
+_TRACKING_USERNAME_ENV_VAR = "MLFLOW_TRACKING_USERNAME"
+_TRACKING_PASSWORD_ENV_VAR = "MLFLOW_TRACKING_PASSWORD"
+_TRACKING_TOKEN_ENV_VAR = "MLFLOW_TRACKING_TOKEN"
+
+# sets verify param of 'requests.request' function
+# see https://requests.readthedocs.io/en/master/api/
+_TRACKING_INSECURE_TLS_ENV_VAR = "MLFLOW_TRACKING_INSECURE_TLS"
+_TRACKING_SERVER_CERT_PATH_ENV_VAR = "MLFLOW_TRACKING_SERVER_CERT_PATH"
+
+# sets cert param of 'requests.request' function
+# see https://requests.readthedocs.io/en/master/api/
+_TRACKING_CLIENT_CERT_PATH_ENV_VAR = "MLFLOW_TRACKING_CLIENT_CERT_PATH"
+
+host_metadata_provider = SageMakerMLflowHostMetadataProvider()
+
+
+def get_host_creds(store_uri) -> rest_utils.MlflowHostCreds:
+    """Build MlflowHostCreds for a SageMaker MLflow endpoint.
+
+    Resolves the store URI (ARN) into a URL via SageMakerMLflowHostMetadataProvider,
+    then returns MlflowHostCreds with auth="arn" to trigger SigV4 signing
+    via the AuthProvider entry point.
+    """
+    host_metadata_provider.set_arn(store_uri)
+
+    return rest_utils.MlflowHostCreds(
+        host=host_metadata_provider.construct_tracking_server_url(),
+        username=os.environ.get(_TRACKING_USERNAME_ENV_VAR),
+        password=os.environ.get(_TRACKING_PASSWORD_ENV_VAR),
+        token=os.environ.get(_TRACKING_TOKEN_ENV_VAR),
+        auth="arn",
+        ignore_tls_verification=os.environ.get(_TRACKING_INSECURE_TLS_ENV_VAR) == "true",
+        client_cert_path=os.environ.get(_TRACKING_CLIENT_CERT_PATH_ENV_VAR),
+        server_cert_path=os.environ.get(_TRACKING_SERVER_CERT_PATH_ENV_VAR),
+    )
