+++ sagemaker_mlflow/mlflow_sagemaker_store.py
@@ -11,29 +11,11 @@
 # ANY KIND, either express or implied. See the License for the specific
 # language governing permissions and limitations under the License.
 
-from mlflow.store.tracking.rest_store import RestStore
-from mlflow.utils import rest_utils
-import os
 from functools import partial
 
-from sagemaker_mlflow.mlflow_sagemaker_helpers import SageMakerMLflowHostMetadataProvider
+from mlflow.store.tracking.rest_store import RestStore
 
-host_metadata_provider = SageMakerMLflowHostMetadataProvider()
-
-# Extra environment variables which take precedence for setting the basic/bearer
-# auth on http requests.
-_TRACKING_USERNAME_ENV_VAR = "MLFLOW_TRACKING_USERNAME"
-_TRACKING_PASSWORD_ENV_VAR = "MLFLOW_TRACKING_PASSWORD"
-_TRACKING_TOKEN_ENV_VAR = "MLFLOW_TRACKING_TOKEN"
-
-# sets verify param of 'requests.request' function
-# see https://requests.readthedocs.io/en/master/api/
-_TRACKING_INSECURE_TLS_ENV_VAR = "MLFLOW_TRACKING_INSECURE_TLS"
-_TRACKING_SERVER_CERT_PATH_ENV_VAR = "MLFLOW_TRACKING_SERVER_CERT_PATH"
-
-# sets cert param of 'requests.request' function
-# see https://requests.readthedocs.io/en/master/api/
-_TRACKING_CLIENT_CERT_PATH_ENV_VAR = "MLFLOW_TRACKING_CLIENT_CERT_PATH"
+from sagemaker_mlflow.host_creds import get_host_creds
 
 
 class MlflowSageMakerStore(RestStore):
@@ -42,20 +24,3 @@ class MlflowSageMakerStore(RestStore):
     def __init__(self, store_uri, artifact_uri):
         self.store_uri = store_uri
         super().__init__(partial(get_host_creds, store_uri))
-
-
-def get_host_creds(store_uri) -> rest_utils.MlflowHostCreds:
-    """Configuring mlflow's client"""
-    host_metadata_provider.set_arn(store_uri)
-
-    return rest_utils.MlflowHostCreds(
-        host=host_metadata_provider.construct_tracking_server_url(),
-        username=os.environ.get(_TRACKING_USERNAME_ENV_VAR),
-        password=os.environ.get(_TRACKING_PASSWORD_ENV_VAR),
-        token=os.environ.get(_TRACKING_TOKEN_ENV_VAR),
-        auth="arn",
-        # aws_sigv4="False",  # Auth provider is used instead for now
-        ignore_tls_verification=os.environ.get(_TRACKING_INSECURE_TLS_ENV_VAR) == "true",
-        client_cert_path=os.environ.get(_TRACKING_CLIENT_CERT_PATH_ENV_VAR),
-        server_cert_path=os.environ.get(_TRACKING_SERVER_CERT_PATH_ENV_VAR),
-    )
